Simplified code by using managed components
