+++ Assets/Scenes/SampleScene/Main.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cbc15530fb66af44bac96c83049f16a4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
