+++ Assets/Scripts/Data/CameraTag.cs
@@ -1,0 +1,10 @@
+﻿using Unity.Entities;
+
+namespace Data
+{
+    [GenerateAuthoringComponent]
+    public struct CameraTag : IComponentData
+    {
+
+    }
+}
