+++ Assets/Scripts/Data/CameraTag.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 57b8c222f08942a989d79c7608ea5f15
+timeCreated: 1648688540
