--- Assets/Scripts/Data/HybridStoreAuthor.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 713e84034e61498f90811c6af60ac21d
-timeCreated: 1648289329
