+++ Assets/Scripts/Data/ResourceStoreAuthor.cs
@@ -31,6 +31,7 @@ namespace Data
     {
         None,
         Food,
-        Key
+        Stone,
+        Water
     }
 }
