--- Assets/Scripts/Systems/GameObjects/HybridStore.cs
@@ -1,10 +1,0 @@
-using UnityEngine;
-using UnityEngine.UIElements;
-
-namespace Systems.GameObjects
-{
-    public class HybridStore : MonoBehaviour
-    {
-        public UIDocument screenDoc;
-    }
-}
