--- Assets/Scripts/Systems/GameObjects/HybridStore.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1c5035cc3a23235488f143df5d9ca704
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
