+++ Assets/Scripts/Systems/PhysicsLockingSystem.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 23ebbbda0fd24ad796c4a09c9a4f03aa
+timeCreated: 1648804396
