+++ Assets/UI Toolkit/UI/Node.uss
@@ -1,3 +1,18 @@
+* {
+    border-color: rgba(0, 0, 0, 0);
+    -unity-font-definition: resource('comic');
+    -unity-font-style: bold;
+    color: white;
+    -unity-text-align: middle-center;
+}
+
+.unity-label {
+    color: rgb(255, 255, 255);
+    -unity-text-outline-color: rgba(0, 0, 0, 255);
+    -unity-text-outline-width: 1px;
+    font-size: 32px;
+}
+
 .unity-button {
     background-color: rgba(102, 102, 119, 0.25);
     font-size: 24px;
@@ -13,22 +28,6 @@
     font-size: 20px;
 }
 
-.jump:hover {
+.unity-button.jump:hover {
     background-color: rgba(0,0,200, 0.6);
-    font-size: 20px;
-}
-
-.unity-label {
-    color: rgb(255, 255, 255);
-    -unity-text-outline-color: rgba(0, 0, 0, 255);
-    -unity-text-outline-width: 1px;
-    font-size: 32px;
-}
-
-* {
-    border-color: rgba(0, 0, 0, 0);
-    -unity-font-definition: resource('comic');
-    -unity-font-style: bold;
-    color: white;
-    -unity-text-align: middle-center;
 }
